+++ HISTORY.rst
@@ -12,6 +12,7 @@ Pending Release
 * Stopped including the 'tests' directory in package
 * Django 1.8 and 1.9 supported.
 * Python 3 support added.
+* Fixed ``setdefault()`` to return the value that was set/found, as per normal dict semantics. Thanks @olevinsky.
 
 1.4.1 (2012-12-04)
 ------------------